+++ src/Handler/Research.hs
@@ -21,7 +21,7 @@ import Research.Tree ( techMap )
 -- | Api to retrieve currently available research
 getApiAvailableResearchR :: Handler Value
 getApiAvailableResearchR = do
-    (_, _, fId) <- apiRequireFaction
+    (_, _, _, fId) <- apiRequireFaction
     available <- runDB $ selectList [ AvailableResearchFactionId ==. fId ] []
     let tech = availableResearchType . entityVal <$> available
     let research = techMap <$> tech
@@ -31,7 +31,7 @@ getApiAvailableResearchR = do
 -- | Api to retrieve all research currently in progress
 getApiCurrentResearchR :: Handler Value
 getApiCurrentResearchR = do
-    (_, _, fId) <- apiRequireFaction
+    (_, _, _, fId) <- apiRequireFaction
     res <- runDB $ loadCurrentResearch fId
     return $ toJSON res
 
@@ -39,7 +39,7 @@ getApiCurrentResearchR = do
 -- | Api to start new research
 postApiCurrentResearchR :: Handler Value
 postApiCurrentResearchR = do
-    (_, _, fId) <- apiRequireFaction
+    (_, _, _, fId) <- apiRequireFaction
     newRes <- requireJsonBody
     available <- runDB $ selectList [ AvailableResearchFactionId ==. fId ] []
     let validRes = validateNewResearch (entityVal <$> available) newRes
@@ -51,7 +51,7 @@ postApiCurrentResearchR = do
 -- | Api to delete research currently in progress
 deleteApiCurrentResearchR :: Handler Value
 deleteApiCurrentResearchR = do
-    (_, _, fId) <- apiRequireFaction
+    (_, _, _, fId) <- apiRequireFaction
     newRes <- requireJsonBody
     current <- runDB $ selectList [ CurrentResearchFactionId ==. fId
                                   , CurrentResearchType ==. (researchType . researchProgressResearch) newRes] []
@@ -65,7 +65,7 @@ deleteApiCurrentResearchR = do
 -- | Api to get current research production
 getApiResearchProductionR :: Handler Value
 getApiResearchProductionR = do
-    (_, _, fId) <- apiRequireFaction
+    (_, _, _, fId) <- apiRequireFaction
     pnbs <- runDB $ factionBuildings fId
     let buildings = pnbs >>= snd
     let total = mconcat $ (researchOutput . entityVal) <$> buildings