+++ src/Model.hs
@@ -14,7 +14,7 @@ import ClassyPrelude.Yesod
 import Database.Persist.Quasi
 import CustomTypes
 import Vehicles.Components
-import People
+import People.Data
 import Research.Data
 
 -- You can define all of your database entities in the entities file.